requirements.txt
requests
pycryptodome
curl_cffi
aiohttp
certifi
browser_cookie3
websockets
js2py
flask
flask-cors
typing-extensions
transformers
PyExecJS
tensorflow
torch
quart
fastapi
uvicorn
python-dotenv
asyncio